pyproject.toml
[tool.poetry]
name = "mini-assessment-engine"
version = "0.1.0"
description = "A mini assessment engine for acad ai"
authors = ["Daniel Adewole <adedaniel504@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
django = "5.2"
djangorestframework = ">=3.16.1,<4.0.0"
drf-spectacular = ">=0.27.0,<1.0.0"
psycopg2-binary = ">=2.9.0,<3.0.0"
python-decouple = ">=3.8,<4.0.0"
dj-database-url = ">=2.1.0,<3.0.0"
scikit-learn = ">=1.6.0,<2.0.0"
openai = {version = ">=1.0.0,<2.0.0", optional = true}
django-filter = "^25.2"


[tool.poetry.extras]
ai-grading = ["openai"]


[tool.poetry.group.dev.dependencies]
ruff = ">=0.8.0,<1.0.0"

[tool.ruff]
# Same as Black
line-length = 88
target-version = "py311"

# Exclude directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*/migrations/*",
    "*.egg-info",
    "build",
    "dist",
]

[tool.ruff.lint]
# Enable Pyflakes (F), pycodestyle (E, W), and isort (I)
select = ["E", "F", "W", "I"]

# Ignore rules that conflict with the formatter
ignore = [
    "E203",  # Whitespace before ':'
    "E501",  # Line too long (handled by formatter)
]

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"

# Like Black, respect magic trailing commas
skip-magic-trailing-comma = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
